Role-based Acces control & permission

+++ SecuringAngularApps.API/Controllers/ProjectsController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using IdentityModel;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -13,7 +14,7 @@ namespace SecuringAngularApps.API.Controllers
 {
     [Produces("application/json")]
     [Route("api/Projects")]
-    [Authorize]
+    [Authorize(Roles = "Admin")]
     public class ProjectsController : Controller
     {
         private readonly ProjectDbContext _context;
@@ -25,9 +26,22 @@ namespace SecuringAngularApps.API.Controllers
 
         // GET: api/Projects
         [HttpGet]
+        //[AllowAnonymous]
         public IEnumerable<Project> GetProjects()
         {
-            return _context.Projects;
+            //var claims = (from c in User.Claims select new { c.Type, c.Value }).ToList();
+            //claims.ForEach(c => Console.WriteLine($"{c.Type}: {c.Value}"));
+            if (User.IsInRole("Admin"))
+            {
+                return _context.Projects;
+            }
+            else
+            {
+                var userId = this.User.FindFirstValue(JwtClaimTypes.Subject);
+                List<int> userProjectIds = _context.UserPermissions.Where(up =>
+                   up.ProjectId.HasValue && up.UserProfileId == userId).Select(up => up.ProjectId.Value).ToList();
+                return _context.Projects.Where(p => userProjectIds.Contains(p.Id));
+            }
         }
 
         // GET: api/Projects/5
@@ -38,7 +52,7 @@ namespace SecuringAngularApps.API.Controllers
             {
                 return BadRequest(ModelState);
             }
-
+            if (!await ProjectEditAccessCheck(id, false) && !User.IsInRole("Admin")) return Forbid();
             var project = await _context.Projects
                 .Include("UserPermissions")
                 .Include("Milestones")
@@ -53,6 +67,7 @@ namespace SecuringAngularApps.API.Controllers
         }
 
         [HttpGet("{id}/Users")]
+        [Authorize(Roles = "Admin")]
         public IActionResult GetProjectUsers([FromRoute] int id)
         {
             var users = from u in _context.UserProfiles.Include("UserPermissions")
@@ -71,12 +86,11 @@ namespace SecuringAngularApps.API.Controllers
             {
                 return BadRequest(ModelState);
             }
-
             if (id != project.Id)
             {
                 return BadRequest();
             }
-
+            if (!await ProjectEditAccessCheck(id, true) && !User.IsInRole("Admin")) return Forbid();
             _context.Entry(project).State = EntityState.Modified;
 
             try
@@ -100,6 +114,7 @@ namespace SecuringAngularApps.API.Controllers
 
         // POST: api/Projects
         [HttpPost]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> PostProject([FromBody] Project project)
         {
             if (!ModelState.IsValid)
@@ -115,6 +130,7 @@ namespace SecuringAngularApps.API.Controllers
 
         // DELETE: api/Projects/5
         [HttpDelete("{id}")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> DeleteProject([FromRoute] int id)
         {
             if (!ModelState.IsValid)
@@ -141,6 +157,7 @@ namespace SecuringAngularApps.API.Controllers
         {
             var item = await _context.Milestones.FirstOrDefaultAsync(m => m.Id == milestone.Id);
             if (item != null) return StatusCode(409);
+            if (!await MilestoneAccessCheck(milestone) && !User.IsInRole("Admin")) return Forbid();
             _context.Milestones.Add(milestone);
             await _context.SaveChangesAsync();
             return CreatedAtAction("GetProject", new { id = milestone.ProjectId }, milestone);
@@ -151,6 +168,7 @@ namespace SecuringAngularApps.API.Controllers
         {
             var item = await _context.Milestones.FirstOrDefaultAsync(m => m.Id == id);
             if (item == null) return NotFound();
+            if (!await MilestoneAccessCheck(item) && !User.IsInRole("Admin")) return Forbid();
             _context.Milestones.Remove(item);
             await _context.SaveChangesAsync();
             return Ok();
@@ -162,10 +180,28 @@ namespace SecuringAngularApps.API.Controllers
             if (milestone.Id != id) return BadRequest();
             var item = await _context.Milestones.FirstOrDefaultAsync(ms => ms.Id == id);
             if (item == null) return NotFound();
+            if (!await MilestoneAccessCheck(item) && !User.IsInRole("Admin")) return Forbid();
             item.MilestoneStatusId = milestone.MilestoneStatusId;
             item.Name = milestone.Name;
             await _context.SaveChangesAsync();
             return Ok(milestone);
+        }
+
+        private async Task<bool> MilestoneAccessCheck(Milestone item)
+        {
+            var userId = this.User.FindFirstValue(JwtClaimTypes.Subject);
+            var perm = await _context.UserPermissions.FirstOrDefaultAsync(up =>
+            up.ProjectId == item.ProjectId &&
+                up.UserProfileId == userId);
+            return (perm != null && perm.Value == "Edit");
+        }
+
+        private async Task<bool> ProjectEditAccessCheck(int projectId, bool edit)
+        {
+            var userId = this.User.FindFirstValue(JwtClaimTypes.Subject);
+            var userAccess = await _context.UserPermissions.FirstOrDefaultAsync(up =>
+                up.ProjectId == projectId && up.UserProfileId == userId);
+            return (userAccess != null && (edit ? userAccess.Value == "Edit" : true));
         }
 
         [HttpGet("MilestoneStatuses")]
@@ -178,6 +214,5 @@ namespace SecuringAngularApps.API.Controllers
         {
             return _context.Projects.Any(e => e.Id == id);
         }
-
     }
 }
